pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-templates-latitude"
version = "0.1.0"
description = "LLM template loader for Latitude - Load prompts from Latitude as LLM templates"
readme = "README.md"
authors = [{name = "Pablo Caro Revuelta", email = "correo@pablocaro.es"}]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "llm>=0.13",
    "httpx>=0.24.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/pcaro/llm-templates-latitude"
Issues = "https://github.com/pcaro/llm-templates-latitude/issues"


[project.entry-points.llm]
templates_latitude = "llm_templates_latitude"

# UV-specific configuration
[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff]
line-length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
